Fix timeout and empty folder errors
the checkDB command timed out (default 30 seconds) on larger databases which left the script in a bad state, unable to check later databases.  Set timeout to 32k seconds or 5.5 hours.

When scanning the subfolders, if no .bak file exists it now skips the run moving on to the next folder.  Can happen if a database is removed the but backup folder isn't removed once the files are cleaned up.